+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Ubuntu 19.04 (disco) support
+- Strip Nginx binary to remove debug symbols
 
 ### Changed